Added 'nxmerge' script
This was in the nxpeaks repository, but it’s a commonly needed
function. The only problem is that it assumes that the TIFF and CBF
modules used by the NeXpy readers have been installed. These are not
listed as a requirement so that they don’t cause installation problems
for  the nexusformat package.