Fixed minor bugs
If an NXfield contained the ‘axes’ attribute, it was not properly
changed when projections summed over an axis. This didn’t show up when
plotting an NXdata group, because plots use the group ‘axes’ attribute
by default, but it would cause an error when trying to plot the NXfield
itself. Also fixes a minor external link bug.