Make NXlinkexternal work with NXFileRemote
If a remote file contains an external link, the link cannot be opened
as an NXFile object. In those case, the nxfile and nxpath properties
need to refer to the main file, not the link. However, I don’t want to
directly check for an NXFileRemote parent in case the pyro directory is
not available, so I need a more indirect check.