Fix bug with copying between files
If an NXfield was too large to be loaded, the data was copied between
files using the h5py group copy function. However, that would fail if a
dataset of the right name already exists in the target file. By adding
the __contains__ function, the path is now checked first and the
dataset replaced if necessary.