Fallback with UnicodeDecodeError
The default Unicode encoding for NeXpy is UTF-8, which is used by HDF5
for variable-length strings. If a file contains text with an alternate
encoding, converting the text to a Unicode string can cause errors.
Since the encoding is not specified anywhere, it is impossible to
guarantee the correct conversion. This is a temporary fallback that
tries ‘latin1’ (or ISO8859-1) encoding, as the most commonly used
alternative.