Change to NXfield nxaxes property
If the axes are not defined by the NXfield’s attributes, the nxaxes
property will try to return the axes defined by the parent group, if
the NXfield is defined as the group’s signal. There is no point in
using the group’s axes if they are intended for another signal NXfield.