Improved handling of singleton and reversed axes
The NXgroup project method has been moved to NXdata, where it is
applicable. It will no longer remove singleton axes from the result if
they were not part of the projection, since that might be an unintended
side effect in some cases. The convert_index function correctly handles
reversed axes.