Improve PyQt5 compatibility
The plot methods need to check to see if they are being run within
NeXpy. If not, they make standard matplotlib calls (by default), which
work in a standard python shell. The previous check entailed indirectly
importing PyQt4, which triggered an error if run within a PyQt5 IPython
qtconsole. This version checks for the existence of the ‘plotview’
variable without importing PyQt unnecessarily.