Add chunks and compression to core memory file
If an NXfield is initialized in slabs, the values are stored in an HDF5
core memory file until it is saved to disk. Since the save just does an
h5py copy, it is important to use chunks, compression, and fill values
(when specified), at the time the core memory dataset is created.