Store selected chunk size
The chunk sizes are stored when they are set in core memory, i.e., when
slab entries are added to an NXfield that is not stored in a file. This
also fixes a Python3 bug when a list of strings is stored. HDF5 doesn’t
seem to recognize a Unicode datatype, e.g. ‘<U5’.