Improve NXobject comparisons
The __eq__ function now checks for object identity before comparing
values. This helps in determining if an object is in a group, for
example, when a comparison of Numpy values raises a ValueError even
when they are the same. This also fixes a Python 3 requirement for
hashable objects to have __hash__ defined when __eq__ is defined.