Refactoring of NXlink code
The previous structure did not allow for groups to be externally
linked. Fixing this required a substantial change in the NXlink code,
to allow external links to be created, even when the file is not yet
present, and to allow them to be read with a distinctive class. The
current version reads files with external links, but writing is not yet
fully debugged.