Fix writing of external links
This has involved extensive modifications to ensure that the external
link is incorporated into an already open NeXus tree, when the tree has
been saved to a file. A new _writeexternal function has been added to
NXFile, replacing the old linkexternal function, and a new NXlink
update function attempts to read newly added external links. This only
works for links added to existing NeXus files because otherwise the
relative file path is unknown. If the link is added to a NeXus tree
that has not been saved, it is left unresolved as an Xlink object.