Fix handling of internal links within external links
If there is an internal link in an HDF5 file that is externally linked,
it could not be resolved because the NXFile _readlink function did not
return an external filename. This also makes sure that the NXgroup
insert function copies the values from the linked field if it is an
external link.