Fix case where relative file path of linked file not correct
A NXlink to an external file may point to a non-existent file, perhaps
because a script is being run in the wrong directory. This was meant to
be trapped in an exception clause, but h5py creates the file in the
current directory without raising an error. Rather than using
error-trapping, the code now checks that both the file and target exist
before trying to read the link.