Resolve links in plotting data
If a link is used in a new NXdata group for plotting, it is necessary
to convert the links within the new group to the parent fields.
Otherwise, it may not be possible to resolve the link target. This also
fixes a problem with axes returned by the ‘nxaxes’ property when a
deprecated axis scheme is used.