Make NXattr and AttrDict more robust
The main change is to always store the raw array stored in the HDF5 file that is accessible with the `nxdata` attribute, but create a new `value` attribute that converts byte strings to strings, and string arrays to a list of strings. In AttrDict, the returned dictionary value calls the latter, which is the most useful in programming, but the `get` function returns the NXattr object itself, providing access to the raw array.