Add IndexError to handled exceptions
If the `axes` attribute exists, but has the wrong shape, this will trigger an IndexError. In principle, this should be raised, but there are so many non-standard NeXus files and customizing exceptions for each one is a major job. Instead, this will enable fallback mechanisms for plotting data without properly defined axes.