Confine creation of core memory file to large datasets
If a value is added using __getitem__, when `_value` is still `None` for a small array, the remaining values can be set to the fill value without needing to create a core memory file. It is still best to make sure that the fill value is defined before adding any data to the NXfield. The default fill value is 0 or a space for string dtypes.