Allow nested 'with' statements
This keeps a count of nested 'with' statements, only releasing the lock and closing the file when all of them are ready to be closed. This provides a mechanism for  combining multiple interactions with the file without requiring a fresh lock for each one. The NXFile lock object is now persistent, with locks turned off by setting timeout to 0.