Add option to load files non-recursively
Only the top level of groups will be loaded if the 'recursive=False' keyword is used to open or load the file.  This speeds up file loading considerably when the files contain many objects. Since other entries are automatically read when they are referenced, the only effect should be an increase in the latency of those references.